- Solved all issues

+++ SFSCollectionMenuDemo/SFSCollectionMenuDemo/SFSViewController.m
@@ -60,7 +60,7 @@
 }
 
 - (void)controller:(SFSCollectionMenuController *)controller didTapButtonAtIndexPath:(NSIndexPath *)indexPath {
-    NSString *text = [NSString stringWithFormat:@"Button %d tapped", indexPath.row];
+    NSString *text = [NSString stringWithFormat:@"Button %ld tapped", (long)indexPath.row];
     self.buttonTappedLabel.text = text;
 }
